feat: integrate daisyUI library
Updates the project to include the daisyUI library, which is a modern CSS framework that provides a set of pre-styled components and utilities that can be used to quickly and easily create beautiful and responsive web applications.

This change will allow us to develop the UI of the application more efficiently, reducing the amount of custom CSS and styling that we need to write ourselves. It will also ensure that our UI is consistent and adheres to best practices for modern web development.